requirements.txt
langchain 
langchain-core
langchain-openai
langchain-community

faiss-cpu
pypdf
tiktoken
python-dotenv
langgraph
fastapi
uvicorn
streamlit
requests
psycopg[binary]
langgraph-checkpoint-postgres